Extract date parsing methods into DateParser
