+++ src/main/java/duke/task/Deadline.java
@@ -1,7 +1,7 @@
 package duke.task;
 
+import duke.DateParser;
 import duke.DukeException;
-import duke.Parser;
 
 public class Deadline extends Task {
     protected String date;
@@ -9,7 +9,7 @@ public class Deadline extends Task {
     public Deadline(String desc, String date) {
         super(desc);
         try {
-            this.date = Parser.parseDateOrDateTimeToString(date);
+            this.date = DateParser.parseDateOrDateTimeToString(date);
         } catch (DukeException e) {
             this.date = date;
         }
